+++ src/components/Claim/Claim.consts.ts
@@ -1,6 +1,5 @@
-import * as interfaces from '@/interfaces'
 import * as consts from '@/constants'
-import { interfaces as peanutInterfaces } from '@squirrel-labs/peanut-sdk'
+import * as interfaces from '@/interfaces'
 import { IOfframpConfirmScreenProps, IOfframpSuccessScreenProps } from '../Offramp/Offramp.consts'
 export type ClaimType = 'claim' | 'claimxchain'
