--- src/components/Create/Link/Input.view.tsx
@@ -1,466 +1,0 @@
-'use client'
-
-import TokenAmountInput from '@/components/Global/TokenAmountInput'
-import TokenSelector from '@/components/Global/TokenSelector/TokenSelector'
-import { useAppKit } from '@reown/appkit/react'
-import { useContext, useEffect, useMemo, useState } from 'react'
-import { useCreateLink } from '../useCreateLink'
-
-import { Button, Card } from '@/components/0_Bruddle'
-import { useToast } from '@/components/0_Bruddle/Toast'
-import FileUploadInput from '@/components/Global/FileUploadInput'
-import FlowHeader from '@/components/Global/FlowHeader'
-import Icon from '@/components/Global/Icon'
-import MoreInfo from '@/components/Global/MoreInfo'
-import { PEANUT_WALLET_CHAIN, PEANUT_WALLET_TOKEN } from '@/constants'
-import * as context from '@/context'
-import { useWalletType } from '@/hooks/useWalletType'
-import { useZeroDev } from '@/hooks/useZeroDev'
-import { useWallet } from '@/hooks/wallet/useWallet'
-import { WalletProviderType } from '@/interfaces'
-import {
-    balanceByToken,
-    ErrorHandler,
-    floorFixed,
-    isNativeCurrency,
-    printableAddress,
-    printableUsdc,
-    validateEnsName,
-} from '@/utils'
-import { interfaces } from '@squirrel-labs/peanut-sdk'
-import { formatEther } from 'viem'
-import * as _consts from '../Create.consts'
-import { isGaslessDepositPossible } from '../Create.utils'
-import * as Sentry from '@sentry/nextjs'
-
-export const CreateLinkInputView = ({
-    onNext,
-    onPrev,
-    tokenValue,
-    setTokenValue,
-    usdValue,
-    setUsdValue,
-    setLinkDetails,
-    setPassword,
-    setGaslessPayload,
-    setGaslessPayloadMessage,
-    setPreparedDepositTxs,
-    setTransactionType,
-    setTransactionCostUSD,
-    setFeeOptions,
-    setEstimatedPoints,
-    attachmentOptions,
-    setAttachmentOptions,
-    createType = 'link', // note: default to link view temporarily
-    recipient,
-    crossChainDetails,
-}: _consts.ICreateScreenProps) => {
-    const {
-        generateLinkDetails,
-        checkUserHasEnoughBalance,
-        generatePassword,
-        makeGaslessDepositPayload,
-        prepareDepositTxs,
-        switchNetwork,
-        estimateGasFee,
-        estimatePoints,
-        prepareDirectSendTx,
-    } = useCreateLink()
-    const {
-        selectedTokenPrice,
-        inputDenomination,
-        selectedChainID,
-        setSelectedChainID,
-        selectedTokenAddress,
-        setSelectedTokenAddress,
-    } = useContext(context.tokenSelectorContext)
-    const { walletType, environmentInfo } = useWalletType()
-    const { handleLogin } = useZeroDev()
-    const toast = useToast()
-
-    const { setLoadingState, loadingState, isLoading } = useContext(context.loadingStateContext)
-    const [errorState, setErrorState] = useState<{
-        showError: boolean
-        errorMessage: string
-    }>({ showError: false, errorMessage: '' })
-    const [currentInputValue, setCurrentInputValue] = useState<string | undefined>(
-        (inputDenomination === 'TOKEN' ? tokenValue : usdValue) ?? ''
-    )
-
-    const { open } = useAppKit()
-
-    const { selectedWallet, signInModal, isConnected, address, isExternalWallet, isPeanutWallet } = useWallet()
-
-    const handleOnNext = async () => {
-        try {
-            if (isLoading || (isConnected && !currentInputValue)) return
-
-            setLoadingState('Loading')
-
-            setErrorState({
-                showError: false,
-                errorMessage: '',
-            })
-
-            // for native token, estimate gas first before checking balance
-            if (isNativeCurrency(selectedTokenAddress)) {
-                setLoadingState('Loading')
-                const linkDetails = generateLinkDetails({
-                    tokenValue: tokenValue,
-                    envInfo: environmentInfo,
-                    walletType: walletType,
-                })
-
-                const password = await generatePassword()
-                const prepareDepositTxsResponse = await prepareDepositTxs({
-                    _linkDetails: linkDetails,
-                    _password: password,
-                })
-
-                if (prepareDepositTxsResponse) {
-                    const { feeOptions } = await estimateGasFee({
-                        chainId: selectedChainID,
-                        preparedTx: prepareDepositTxsResponse.unsignedTxs[0],
-                    })
-
-                    // calculate gas amount in native token
-                    const gasLimit = BigInt(feeOptions.gasLimit)
-                    const gasPrice = BigInt(feeOptions.maxFeePerGas || feeOptions.gasPrice)
-                    const maxGasAmount = Number(formatEther(gasLimit * gasPrice))
-
-                    setLoadingState('Loading')
-                    await checkUserHasEnoughBalance({
-                        tokenValue: tokenValue,
-                        gasAmount: maxGasAmount,
-                    })
-                }
-            } else {
-                setLoadingState('Loading')
-                await checkUserHasEnoughBalance({ tokenValue: tokenValue })
-            }
-
-            setLoadingState('Generating details')
-
-            const linkDetails = generateLinkDetails({
-                tokenValue: tokenValue,
-                envInfo: environmentInfo,
-                walletType: walletType,
-            })
-            setLinkDetails(linkDetails)
-
-            if (createType !== 'direct') {
-                const password = await generatePassword()
-                setPassword(password)
-
-                setLoadingState('Preparing transaction')
-
-                let prepareDepositTxsResponse: interfaces.IPrepareDepositTxsResponse | undefined
-                const _isGaslessDepositPossible = isGaslessDepositPossible({
-                    chainId: selectedChainID,
-                    tokenAddress: selectedTokenAddress,
-                })
-                if (_isGaslessDepositPossible && selectedWallet?.walletProviderType !== WalletProviderType.PEANUT) {
-                    // PW userops are marked as 'not-gasless' in this flow, since
-                    // they will become gasless via the paymaster
-                    setTransactionType('gasless')
-
-                    const makeGaslessDepositResponse = await makeGaslessDepositPayload({
-                        _linkDetails: linkDetails,
-                        _password: password,
-                    })
-
-                    if (
-                        !makeGaslessDepositResponse ||
-                        !makeGaslessDepositResponse.payload ||
-                        !makeGaslessDepositResponse.message
-                    )
-                        return
-                    setGaslessPayload(makeGaslessDepositResponse.payload)
-                    setGaslessPayloadMessage(makeGaslessDepositResponse.message)
-
-                    setFeeOptions(undefined)
-                    setTransactionCostUSD(0)
-                } else {
-                    setTransactionType('not-gasless')
-
-                    /*
-                     * DEV BLOCK: Code to reset token approvals for testing
-                     *
-                     * This code can be uncommented to force token approval to 0
-                     * in order to test the approval flow:
-                     *
-                     * let chainId = linkDetails.chainId
-                     * let tokenAddress = linkDetails.tokenAddress!
-                     * let provider = undefined
-                     * let spenderAddress = undefined
-                     *
-                     * let contractVersion = getLatestContractVersion({
-                     *   chainId: linkDetails.chainId,
-                     *   type: 'normal'
-                     * })
-                     *
-                     * const defaultProvider = provider || (await getDefaultProvider(chainId))
-                     * const tokenContract = new ethers.Contract(tokenAddress, ERC20_ABI, defaultProvider)
-                     *
-                     * const _PEANUT_CONTRACTS = PEANUT_CONTRACTS as {
-                     *   [chainId: string]: { [contractVersion: string]: string }
-                     * }
-                     * const spender = spenderAddress || (
-                     *   _PEANUT_CONTRACTS[chainId] && _PEANUT_CONTRACTS[chainId][contractVersion]
-                     * )
-                     *
-                     * const tx = await tokenContract.populateTransaction.approve(spender, BigInt(0))
-                     * const peanutTxs: any = {unsignedTxs: [ethersV5ToPeanutTx(tx)]}
-                     *
-                     * prepareDepositTxsResponse = peanutTxs
-                     */
-
-                    prepareDepositTxsResponse = await prepareDepositTxs({
-                        _linkDetails: linkDetails,
-                        _password: password,
-                    })
-
-                    setPreparedDepositTxs(prepareDepositTxsResponse)
-
-                    let _feeOptions = undefined
-
-                    try {
-                        const { feeOptions, transactionCostUSD } = await estimateGasFee({
-                            chainId: selectedChainID,
-                            preparedTx: prepareDepositTxsResponse?.unsignedTxs[0],
-                        })
-
-                        _feeOptions = feeOptions
-                        setFeeOptions(feeOptions)
-                        setTransactionCostUSD(transactionCostUSD)
-                    } catch (error) {
-                        console.error(error)
-                        setFeeOptions(undefined)
-                        setTransactionCostUSD(undefined)
-                        Sentry.captureException(error)
-                    }
-                }
-
-                const estimatedPoints = await estimatePoints({
-                    chainId: selectedChainID,
-                    address: address ?? '',
-                    amountUSD: parseFloat(usdValue ?? '0'),
-                    preparedTx: _isGaslessDepositPossible
-                        ? undefined
-                        : prepareDepositTxsResponse &&
-                          prepareDepositTxsResponse?.unsignedTxs[prepareDepositTxsResponse?.unsignedTxs.length - 1],
-                    actionType: 'CREATE',
-                })
-
-                if (estimatedPoints) setEstimatedPoints(estimatedPoints)
-                else setEstimatedPoints(0)
-            } else {
-                const preparedTxs: interfaces.IPrepareDepositTxsResponse = {
-                    unsignedTxs: [
-                        {
-                            ...prepareDirectSendTx({
-                                tokenValue: tokenValue ?? '',
-                                recipient: recipient.address ?? '',
-                                tokenAddress: selectedTokenAddress,
-                                tokenDecimals: linkDetails.tokenDecimals,
-                            }),
-                        },
-                    ],
-                }
-
-                setPreparedDepositTxs(preparedTxs)
-
-                const estimatedPoints = await estimatePoints({
-                    chainId: selectedChainID,
-                    address: address ?? '',
-                    amountUSD: parseFloat(usdValue ?? '0'),
-                    preparedTx: preparedTxs?.unsignedTxs[preparedTxs?.unsignedTxs.length - 1],
-                    actionType: 'TRANSFER',
-                })
-
-                if (estimatedPoints) setEstimatedPoints(estimatedPoints)
-                else setEstimatedPoints(0)
-                try {
-                    const { feeOptions, transactionCostUSD } = await estimateGasFee({
-                        chainId: selectedChainID,
-                        preparedTx: preparedTxs?.unsignedTxs[0],
-                    })
-                    setFeeOptions(feeOptions)
-                    setTransactionCostUSD(transactionCostUSD)
-                } catch (error) {
-                    console.error(error)
-                    setFeeOptions(undefined)
-                    setTransactionCostUSD(undefined)
-                    Sentry.captureException(error)
-                }
-            }
-            await switchNetwork(selectedChainID)
-            onNext()
-        } catch (error) {
-            const errorString = ErrorHandler(error)
-            setErrorState({
-                showError: true,
-                errorMessage: errorString,
-            })
-            Sentry.captureException(error)
-        } finally {
-            setLoadingState('Idle')
-        }
-    }
-
-    const maxValue = useMemo(() => {
-        if (!selectedWallet?.balances) {
-            return selectedWallet?.balance ? printableUsdc(selectedWallet.balance) : ''
-        }
-        const balance = balanceByToken(selectedWallet.balances, selectedChainID, selectedTokenAddress)
-        if (!balance) return ''
-        // 6 decimal places, prettier
-        return floorFixed(balance.amount, 6)
-    }, [selectedChainID, selectedTokenAddress, selectedWallet?.balances, selectedWallet?.balance])
-
-    useEffect(() => {
-        if (!currentInputValue) return
-        if (inputDenomination === 'TOKEN') {
-            setTokenValue(currentInputValue)
-            if (selectedTokenPrice) {
-                setUsdValue((parseFloat(currentInputValue) * selectedTokenPrice).toString())
-            }
-        } else if (inputDenomination === 'USD') {
-            setUsdValue(currentInputValue)
-            if (selectedTokenPrice) {
-                setTokenValue((parseFloat(currentInputValue) / selectedTokenPrice).toString())
-            }
-        }
-    }, [currentInputValue, inputDenomination, selectedTokenPrice])
-
-    useEffect(() => {
-        if (isPeanutWallet) {
-            setSelectedChainID(PEANUT_WALLET_CHAIN.id.toString())
-            setSelectedTokenAddress(PEANUT_WALLET_TOKEN)
-        }
-    }, [isPeanutWallet])
-
-    return (
-        <>
-            <FlowHeader />
-            <Card className="shadow-none sm:shadow-primary-4">
-                <Card.Header>
-                    <Card.Title style={{ display: '-webkit-box', WebkitLineClamp: '2', WebkitBoxOrient: 'vertical' }}>
-                        {' '}
-                        {createType === 'link'
-                            ? 'Send to Anyone'
-                            : createType === 'direct'
-                              ? `Send to ${validateEnsName(recipient.name) ? recipient.name : printableAddress(recipient.address ?? '')}`
-                              : `Send to ${recipient.name}`}
-                    </Card.Title>
-                    <Card.Description>
-                        {createType === 'link' &&
-                            "Send funds without needing the recipient's wallet address. They'll get a link to claim in their preferred token and blockchain."}
-                        {createType === 'email_link' &&
-                            `You will send an email to ${recipient.name ?? recipient.address} containing a link. They will be able to claim the funds in any token on any chain from the link.`}
-                        {createType === 'sms_link' &&
-                            `You will send a text message to ${recipient.name ?? recipient.address} containing a link. They will be able to claim the funds in any token on any chain from the link.`}
-                        {createType === 'direct' &&
-                            `You will do a direct blockchain transaction to ${recipient.name ?? recipient.address}. Ensure the recipient address is correct, else the funds might be lost.`}
-                    </Card.Description>
-                </Card.Header>
-                <Card.Content className="flex flex-col gap-4">
-                    <div className="flex flex-col gap-4">
-                        <TokenAmountInput
-                            className="w-full"
-                            tokenValue={currentInputValue}
-                            maxValue={maxValue}
-                            setTokenValue={setCurrentInputValue}
-                            onSubmit={() => {
-                                if (!isConnected) signInModal.open()
-                                else handleOnNext()
-                            }}
-                        />
-                        {isExternalWallet && (
-                            <>
-                                <TokenSelector classNameButton="w-full" />
-                                {selectedWallet!.balances!.length === 0 && (
-                                    <div
-                                        onClick={() => {
-                                            open()
-                                        }}
-                                        className="cursor-pointer text-h9 underline"
-                                    >
-                                        ( Buy Tokens )
-                                    </div>
-                                )}
-                            </>
-                        )}
-                        {(createType === 'link' || createType === 'email_link' || createType === 'sms_link') && (
-                            <FileUploadInput
-                                attachmentOptions={attachmentOptions}
-                                setAttachmentOptions={setAttachmentOptions}
-                            />
-                        )}
-                    </div>
-                    <div className="mb-4 flex flex-col gap-4 sm:flex-row-reverse">
-                        <Button
-                            onClick={() => {
-                                if (!isConnected) {
-                                    if (isPeanutWallet) {
-                                        setLoadingState('Logging in')
-                                        handleLogin()
-                                            .then(() => {
-                                                handleOnNext()
-                                            })
-                                            .catch((error) => {
-                                                Sentry.captureException(error)
-                                                toast.error('Error logging in')
-                                            })
-                                            .finally(() => {
-                                                setLoadingState('Idle')
-                                            })
-                                    } else {
-                                        signInModal.open()
-                                    }
-                                } else {
-                                    handleOnNext()
-                                }
-                            }}
-                            loading={isLoading}
-                            disabled={isLoading || !currentInputValue}
-                        >
-                            {!isConnected && !isPeanutWallet ? 'Connect Wallet' : isLoading ? loadingState : 'Confirm'}
-                        </Button>
-                        {/* <Button variant="stroke" onClick={onPrev} disabled={isLoading}>
-                        Go Back
-                    </Button> */}
-                    </div>
-                    {errorState.showError && (
-                        <div className="text-start">
-                            <label className=" text-h8 font-normal text-red ">{errorState.errorMessage}</label>
-                        </div>
-                    )}
-                    {!crossChainDetails.find(
-                        (chain: any) => chain.chainId.toString() === selectedChainID.toString()
-                    ) && (
-                        <span className=" text-start text-h8 font-normal">
-                            <Icon name="warning" className="-mt-0.5" /> This chain does not support cross-chain
-                            claiming.
-                        </span>
-                    )}
-
-                    <span className="flex flex-row items-center justify-start gap-1 text-h8">
-                        Learn about Peanut cashout
-                        <MoreInfo
-                            text={
-                                <>
-                                    You can use Peanut to cash out your funds directly to your bank account! (US and EU
-                                    only)<br></br>{' '}
-                                    <a href="/cashout" className="hover:text-primary underline">
-                                        Learn more →
-                                    </a>
-                                </>
-                            }
-                        />
-                    </span>
-                </Card.Content>
-            </Card>
-        </>
-    )
-}
