--- src/components/Create/Link/index.ts
@@ -1,4 +1,0 @@
-export * from './Initial.view'
-export * from './Confirm.view'
-export * from './Success.view'
-export * from './Input.view'
