--- src/components/Create/index.ts
@@ -1,1 +1,0 @@
-export * from './Create'
