+++ src/redux/slices/payment-slice.ts
@@ -1,9 +1,9 @@
-import { IAttachmentOptions } from '@/components/Create/Create.consts'
 import { ParsedURL } from '@/lib/url-parser/types/payment'
 import { PaymentCreationResponse, TCharge, TRequestChargeResponse, TRequestResponse } from '@/services/services.types'
 import { createSlice, PayloadAction } from '@reduxjs/toolkit'
 import { PAYMENT_SLICE } from '../constants'
 import { IPaymentState, TPaymentView } from '../types/payment.types'
+import { IAttachmentOptions } from '../types/send-flow.types'
 
 const initialState: IPaymentState = {
     currentView: 'INITIAL',
