+++ src/redux/types/payment.types.ts
@@ -1,6 +1,6 @@
-import { IAttachmentOptions } from '@/components/Create/Create.consts'
 import { ParsedURL } from '@/lib/url-parser/types/payment'
 import { PaymentCreationResponse, TCharge, TRequestChargeResponse, TRequestResponse } from '@/services/services.types'
+import { IAttachmentOptions } from './send-flow.types'
 
 export type TPaymentView = 'INITIAL' | 'CONFIRM' | 'STATUS'
